Feature/templates pre load query (#39)
* Templates - Added functionality for pre load query

In HTML templates you can now setup a query that will be executed at the start of the rendering of that template. If that query returns any results, those results will be stored in the HttpContext. Then those results can be used in any template or query of any dynamic component on that template. This can be used to gather some information beforehand, that you need in multiple components (such as the ID of a product, based on the URL).

* Templates - Added table "wiser_dynamic_content_publish_log" for logging publish actions of dynamic components.

* Created version 1.2.10