Added DatabaseTimeZone property to GclSettings
if this has a value, the time zone will always be set to the connection to database, before executing any query. This is to make sure that times saved in the database will be returned in the time zone we want, instead of the time zone that they are saved as.