WiserItemsService - Changed the update function
It can no longer be used to change the entity type of an item. It's really weird to change the entity type of an item, that should almost never happen. If you do really need to change the entity type, use the function "ChangeEntityTypeAsync" instead of "UpdateAsync".

This is a workaround for a bug in Wiser; In some cases Wiser sends the wrong entity type to the update function, which caused items to suddenly have a different entity type. I was not able to reproduce this problem myself, so made this workaround to still fix that problem.