Login required support for templates (#113)
* Login required support for templates

The "login required" option for templates is now supported. The templates service will return an empty template if a login is required, but no login is found (or if no HttpContext is available). The templates controller will redirect the user to whatever is set in the "login redirect URL" field, or return a 401 status (Unauthorized) if no redirect URL is set. Query and partial templates always return a 401 status if a login is required and no active login is found.

* Fixed error after previous merge conflict.