Fixed WebPages cache issues (#219)
* Made WebPages cache clearable

The WebPages cache couldn't be cleared because the caching functions
didn't assign a cancellation token to the cache entries. Also updated
the GetOrAddAsync calls to not use delegates (so it's consistent with
the rest of the GCL).

* Fixed cache key for retrieving page contents

The generated cache key when retrieving page data didn't take request
variables and extra data into account.