Improve the query for the path contains functionality in webpages.
When a webpage is in two folders and those folders have a shared part the order was deciding in which webpage was selected.

For example a webpage named "information" in the folders "about products" and "about" would give the webpage in the "about products" folder instead of "about" when the path must contain "about".

By adding an empty string before and after the path it will create a "/" before and after so the path contains can be set with "/about/" to refer to the correct folder.