Configurator sub steps placement
Changed the way the sub steps can be placed in the HTML by allowing only
certain sub steps (by either ID or name) to be placed in specific parts
of the HTML, by using pipes to separate the IDs or names of the
substeps, like so: `{substeps|789|123}` or `{substeps|Step 4|Step 2}`.
The regular `{substeps}` is also still available.