Divided coupon discount (#274)
* Started development for divided coupon discount

This is a new setting that will divide a coupon's discount over all
products instead of having one discount value. This also adds support
for a new setting that allows products to be excluded from coupon
discounts through a query.

* Finished the divided coupon discount functionality

Also added a lot of missing comments and fixed some compiler warnings.