Changed how the languages middleware works
The languages middleware used to check if the page needed the language code to be stored in the session by checking the current URL against a regex. However, this caused many timeouts in the regex when many requests were made at the same time. On top of that, it's probably slower to use the regex instead of just setting the language session, so the entire check has been removed.