Removed languages middleware entirely
The languages middleware has been removed and an additional replacements has been added for the language code: "MlJclLanguageCode" (to ensure the legacy support stays intact).

The reason the middleware is removed:
After some internal discussion it was decided that the languages middleware is not needed anymore, as dependency injection will always ensure the languages service is available for the language code replacements.