Changed how the languages middleware works (#342)
* Changed how the languages middleware works

The languages middleware used to check if the page needed the language code to be stored in the session by checking the current URL against a regex. However, this caused many timeouts in the regex when many requests were made at the same time. On top of that, it's probably slower to use the regex instead of just setting the language session, so the entire check has been removed.

* Removed languages middleware entirely

The languages middleware has been removed and an additional replacements has been added for the language code: "MlJclLanguageCode" (to ensure the legacy support stays intact).

The reason the middleware is removed:
After some internal discussion it was decided that the languages middleware is not needed anymore, as dependency injection will always ensure the languages service is available for the language code replacements.