Configurator Vue mode (#372)
* Started working on Configurator Vue mode

This mode is intended to be used with the new Vue Configurator library.

* Continued working on Configurator Vue mode

Started work on the "GetVueConfiguratorData" functions.

* Started to work on loading option data

* Configurator - Vue mode development

Step options are now loaded. Also fixed some bugs with loading the step data.

* Configurator - Vue mode development

Added a second version of "ReplaceConfiguratorItemsAsync" for the Vue mode. Also added "CurrentValue" and "CurrentValueDisplayName" to the "VueStepDataModel" class.

* Configurator - Vue mode development

Fixed some issues with the generated HTML and with loading the step options.

* Added required property

The generated "<step>" elements now have the "is-required" prop set.

* Vue configurator development

- Updated "<step>" and "<step-option>" rendering to reflect changes made to the library.
- Steps can now have a regular expression set for validation purposes.
- Required steps can now have conditions that have to met before the step is considered as required.

* Extended step model

Added settings error messages for various validation errors.

* Updated code for legibility

Updated the stringbuilders that place the "<step>" tags for better legibility. Also removed the ":is-required" attribute as it's no longer needed.

* Updated summary HTML

The summary HTML when in Vue mode doesn't get replacements, as it only functions as a template for Vue.