Configurator Vue mode (#386)
* Started working on Configurator Vue mode

This mode is intended to be used with the new Vue Configurator library.

* Continued working on Configurator Vue mode

Started work on the "GetVueConfiguratorData" functions.

* Started to work on loading option data

* Configurator - Vue mode development

Step options are now loaded. Also fixed some bugs with loading the step data.

* Configurator - Vue mode development

Added a second version of "ReplaceConfiguratorItemsAsync" for the Vue mode. Also added "CurrentValue" and "CurrentValueDisplayName" to the "VueStepDataModel" class.

* Configurator - Vue mode development

Fixed some issues with the generated HTML and with loading the step options.

* Added required property

The generated "<step>" elements now have the "is-required" prop set.

* Vue configurator development

- Updated "<step>" and "<step-option>" rendering to reflect changes made to the library.
- Steps can now have a regular expression set for validation purposes.
- Required steps can now have conditions that have to met before the step is considered as required.

* Extended step model

Added settings error messages for various validation errors.

* Updated code for legibility

Updated the stringbuilders that place the "<step>" tags for better legibility. Also removed the ":is-required" attribute as it's no longer needed.

* Updated summary HTML

The summary HTML when in Vue mode doesn't get replacements, as it only functions as a template for Vue.

* Configurator changes

- Added URL regex support.
- Added step display name.
- Removed legacy mode as it was made for the Wiser 1 data model, which is not supported anymore.
- Removed JJL classes from the Vue settings model.
- Made ComponentMode invisible in Wiser.
- Removed "partial" class modifier in the ConfiguratorsService class.

* Vue configurator development

- Separated Vue HTML rendering from Default HTML rendering.
- Added support for extra data for every step.
- Added "available" property which Vue can use to completely hide a step.
- Some other various fixes and changes.

* Fixed summary template

Fixed summary template not being loaded into the "container" template set in the configurator component.

* Updated code based on PR review