Reuse responses if the expected result is the same. (#419)
When the endpoint and body are the same during the loading of multiple steps the responses will be reused. This way less calls need to be made to an external API. Since the data hasn't changed during loading the response should be the same if the endpoint and body are the same.