Fixed inconsistency in "Handle request" content option (#455)
The "HandleRequest" setting for the repeater and pagination dynamic component mentions that it impacts the TEMPLATES of that component, however the pagination code also uses it for the QUERY of the component. The repeater query code does not. This change should fix that and make the setting consistent and comply with the actual description.