Fixed issue with document store
When the document store connection was made, it used the normal MySQL connection string as a fallback if one for the document store was not provided, but this causes various issues. Now, the document store will not attempt to create a connection if a specific document store connection string is not provided, and various checks and logging have been put in place to make sure no errors will arise when the service gets created.