Fixed a caching issue with page widgets
When the page widgets got loaded, they were added to the list of global page widgets, but that caused the cached version of the global page widgets to be populated with the page widgets, causing every page to suddenly start loading that specific page widget. This branch fixes that by always creating a new list and populating it with both the global page widgets and the specific page widgets.