Fix the save history setting on entity level
The GCL did not read this setting from the database and didn't use it for create and update. Deleted used it but was always false since the value was not read.

This change will read the value for the save history setting and add it to the create and update calls.