Fix the save history setting on entity level (#580)
* Fix the save history setting on entity level

The GCL did not read this setting from the database and didn't use it for create and update. Deleted used it but was always false since the value was not read.

This change will read the value for the save history setting and add it to the create and update calls.

* Fix save_history default value

When no entity type ws provided the value would be a DBNull and unable to convert to bool. When none is provided mark it as to save history.