Allow CSS and JS templates in partial templates
Partials templates can now be loaded with CSS and JS templates if they're linked directly to the partial template. If there are CSS and/or JS templates linked to the partial HTML template, a basic view is generated. If no CSS and JS templates are linked, only the partial template's HTML will be returned.