Feature/xss security improvements (#752)
* Improve replacements from request data.

Always strip HTML of user provided data in the request, e.g. querystring and cookies.

* Fixed RXSS vulnerability in Pagination component

An RXSS vulnerability in the Pagination component allowed hackers to
execute scripts through the query string of a page that had the
Pagination component on it. This fix prevents that by using the
QueryString.Create function to rebuild the query string.

---------

Co-authored-by: Hans <35447467+HansJuice@users.noreply.github.com>