Update for "unsafe source" replacements
A small update to the "unsafe source" functionality for the replacements
so it knows what the exact source is for additional control over what
kind of protection should be added (like protection against XSS). For
now, only the HttpRequest is added as an unsafe source, which means that
the values will be URL-encoded.