Changed how we load Evo PDF. We load it with PrivateAssets=true now, which means that it won't be automatically referenced in projects that load the GCL.
This should fix the problem that we'd always only load Evo PDF for Linux, even when projects run on Windows (because the GCL gets built on a Linux machine). This has the added benefits that we now only need to load Evo PDF when projects actually use that functionality, which saves a little loading time.