Fix easy objects from being loaded with no key (#785)
Easy objects were loaded with keys that were `null` to build the cache. If an invalid key was present the GCL would crash and prevent any page from being loaded.

During branch merges this can occur if an object has been added on both the branch as main database before the merge preventing the system to write the key to the main database.