Fix for entity type always being required (#790)
* Fix for entity type always being required

Updating an item with a custom query didn't work anymore due to the
entity type always being required. Setting an entity type, however,
causes the custom query to not be used. This checks if entity type is
set before deciding to use it and falls back to the item's entity type
if it isn't set.

* Removed excess white space in a query

---------

Co-authored-by: Hans <35447467+HansJuice@users.noreply.github.com>