Made Amazon S3 service use file cache (#803)
The Amazon S3 service saves the downloaded objects to the file system,
but didn't use the FileCacheService causing exceptions when multiple
requests were attempting to download the same object. The
FileCacheService is now used.

Also refactored the FileCacheService so it can work with streams as well
and uses SemaphoreSlim to create locks to create thread safety.

Co-authored-by: Hans <35447467+HansJuice@users.noreply.github.com>