Refactor string value conversion in replacements service
Replaced `Convert.ToString` with the more type-safe `jValue.Value<string>()` method. This fixes issues we had with converting decimals to strings for example, because the `ToString` method would use the culture of the PC/server, which caused the problem that we could not convert it back to a decimal again later.