Extend check to include if the transaction was started outside the function
When an exception was thrown handling a query while deleting an item, it would only check if no new transaction needed to be made before performing a rollback. Since the transaction was already active the method didn't start the transaction itself and should not close it. This is done everywhere else in the class.

This caused problems that a rollback was performed but the above lying logic was still thinking a transaction was present and started to execute queries outside of a transaction.